更改了entity内的类熟悉

+++ mall/src/main/java/com/bs/mall/entity/User.java
@@ -33,7 +33,7 @@ public class User {
     /**
      * 用户性别
      */
-    private Integer userGender;
+    private Byte userGender;
 
     /**
      * 用户电话号码
